# Improved performance of `mars.learn.metrics.{roc_curve, roc_auc_score}`