<!--
Thank you for your contribution!

Please review https://github.com/mars-project/mars/blob/master/CONTRIBUTING.rst before opening a pull request.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

This PR improved performance of `mars.learn.metrics.{roc_curve, roc_auc_score}`

### Details

- [x] Support `tileable.cache = True` to cache intermediate tileables.
- [x] Warning when duplicated execution is submitted by enabling `options.warn_duplicated_execution`.
- [x] Add ASV benchmark for `roc_curve` & `roc_auc_score`.
- [x] Improve `roc_curve` & `roc_auc_score` via eliminating duplicated execution.

### Improvements

#### 1. Cache intermediate tileables.

Some tileables were executed multiple times before, now we try to cache them.

#### 2. Optimize by implementing differently

Some optimization are applied including:

##### Optimize `mt.diff` by rechunking better
##### Perform fancy indexing only once

e.g. 

```
t1[ind], t2[ind]
```

now, we do

```
mt.stack([t1, t2])[:, ind]
```

##### Use DataFrame sort to replace tensor argsort + fancy indexes

e.g. 

```
ind = mt.argsort(t2)[::-1]
t1[ind], t2[ind]
```

into

```
df = md.DataFrame({0: t1, 1: t2})
df = df.sort_values(t2, ascending=False)
df[0].to_tensor(), df[1].to_tensor()
```

### Table

| | Master | Improvement 1 |  Improvement 2 |
| --- | --- | --- | --- |
| roc_curve (1 chunk) | 3.04±0.03s | 2.65±0.08s | 2.59±0.01s |
| roc_curve (n chunk) | 18.6±0.05s | 14.9±0.06s | 6.77±0.07s |
| roc_auc_score (1 chunk) | 3.28±0.01s | 2.49±0.01s | 2.43±0.02s |
| roc_auc_score (n chunk) | 17.4±0.1s | 13.2±0s | 6.19±0.01s |

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this change? -->
Fixes #xxxx

## Check code requirements

- [x] tests added / passed (if needed)
- [x] Ensure all linting tests pass, see [here](https://docs.pymars.org/en/latest/development/contributing.html#check-code-styles) for how to run them